refactor the gridProduction and griCompletion into components
